streamlit/requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-chroma
langchain-google-genai
python-dotenv
chromadb
